update tests from branch origin/dm-fallout-tests.

+++ modules/ftm/tests/test_memory.py
@@ -6,12 +6,18 @@ class UnitTestMemory(dm_testbench.DmTestbench):
 
   def setUp(self):
     super().setUp()
-    if self.threadQuantity == 8:
-      self.file1 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_150.dot'
-      self.file2 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_10b.dot'
+    if self.cpuQuantity == 3:
+      self.file1 = 'groups_3_nonDefaultPatterns_9_blocksPerPattern_150.dot'
+      if self.threadQuantity == 8:
+        self.file2 = 'groups_3_nonDefaultPatterns_9_blocksPerPattern_10b.dot'
+      else:
+        self.file2 = 'groups_3_nonDefaultPatterns_9_blocksPerPattern_10c.dot'
     else:
       self.file1 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_150.dot'
-      self.file2 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_10c.dot'
+      if self.threadQuantity == 8:
+        self.file2 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_10b.dot'
+      else:
+        self.file2 = 'groups_4_nonDefaultPatterns_9_blocksPerPattern_10c.dot'
 
   def test_memory_cpu0(self):
     """Test for CPU 0. Add the schedule, add a second schedule: OK. When
@@ -60,9 +66,12 @@ class UnitTestMemoryFull(dm_testbench.DmTestbench):
 
   def setUp(self):
     super().setUp();
+    self.lastCpu = self.cpuQuantity - 1
     if self.threadQuantity == 8:
       self.maxNodes = 1868
       self.maxNodesCpu3 = 1667
+      if self.cpuQuantity == 3:
+        self.maxNodesCpu3 = 1719
     else:
       self.maxNodes = 1839
       self.maxNodesCpu3 = 1634
@@ -111,13 +120,14 @@ class UnitTestMemoryFull(dm_testbench.DmTestbench):
     self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
         fileName), [0], linesCout=0, linesCerr=0)
     self.deleteFile(fileName)
-    fileName = self.schedulesFolder + 'memory_full_4cpuOK_cpu2.dot'
-    self.generate_schedule(fileName, self.maxNodes, 2)
-    self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
-        fileName), [0], linesCout=0, linesCerr=0)
-    self.deleteFile(fileName)
+    if self.cpuQuantity == 4:
+      fileName = self.schedulesFolder + 'memory_full_4cpuOK_cpu2.dot'
+      self.generate_schedule(fileName, self.maxNodes, 2)
+      self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
+          fileName), [0], linesCout=0, linesCerr=0)
+      self.deleteFile(fileName)
     fileName = self.schedulesFolder + 'memory_full_4cpuOK_cpu3.dot'
-    self.generate_schedule(fileName, self.maxNodesCpu3, 3)
+    self.generate_schedule(fileName, self.maxNodesCpu3, self.lastCpu)
     self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
         fileName), [0], linesCout=0, linesCerr=0)
     self.deleteFile(fileName)
@@ -134,13 +144,14 @@ class UnitTestMemoryFull(dm_testbench.DmTestbench):
     self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
         fileName), [0], linesCout=0, linesCerr=0)
     self.deleteFile(fileName)
-    fileName = self.schedulesFolder + 'memory_full_4cpuFail_cpu2.dot'
-    self.generate_schedule(fileName, self.maxNodes, 2)
-    self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
-        fileName), [0], linesCout=0, linesCerr=0)
-    self.deleteFile(fileName)
+    if self.cpuQuantity == 4:
+      fileName = self.schedulesFolder + 'memory_full_4cpuFail_cpu2.dot'
+      self.generate_schedule(fileName, self.maxNodes, 2)
+      self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
+          fileName), [0], linesCout=0, linesCerr=0)
+      self.deleteFile(fileName)
     fileName = self.schedulesFolder + 'memory_full_4cpuFail_cpu3.dot'
-    self.generate_schedule(fileName, self.maxNodesCpu3 + 1, 3)
+    self.generate_schedule(fileName, self.maxNodesCpu3 + 1, self.lastCpu)
     self.startAndCheckSubprocess((self.binaryDmSched, self.datamaster, 'add',
         fileName), [250], linesCout=2, linesCerr=3)
     self.deleteFile(fileName)
